# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "SAID.SALUD API",
    "version": "1.0.0",
    "description": "Documentación técnica de la API REST para el sistema de gestión clínica SAID.SALUD.",
    "contact": {
      "name": "Soporte SAID.SALUD"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000/api/v1",
      "description": "Servidor de Desarrollo Local"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "User": {
        "type": "object",
        "required": [
          "email",
          "role"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "ID autogenerado"
          },
          "email": {
            "type": "string",
            "example": "admin@saidsalud.com"
          },
          "role": {
            "type": "string",
            "enum": [
              "admin",
              "nurse"
            ],
            "example": "admin"
          },
          "nurse_id": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "is_active": {
            "type": "boolean",
            "example": true
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/users": {
      "get": {
        "summary": "Obtener todos los usuarios",
        "tags": [
          "Users"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de usuarios",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "success"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/User"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "No autorizado (Token faltante o inválido)"
          },
          "403": {
            "description": "Prohibido (No tiene permisos de administrador)"
          }
        }
      },
      "post": {
        "summary": "Crear un nuevo usuario",
        "tags": [
          "Users"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "email",
                  "password",
                  "role"
                ],
                "properties": {
                  "email": {
                    "type": "string",
                    "example": "enfermero@saidsalud.com"
                  },
                  "password": {
                    "type": "string",
                    "example": 123456
                  },
                  "role": {
                    "type": "string",
                    "enum": [
                      "admin",
                      "nurse"
                    ],
                    "example": "nurse"
                  },
                  "nurse_id": {
                    "type": "string",
                    "format": "uuid",
                    "example": "uuid-v4"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Usuario creado exitosamente"
          },
          "401": {
            "description": "No autorizado"
          }
        }
      }
    },
    "/users/{id}": {
      "get": {
        "summary": "Obtener usuario por ID",
        "tags": [
          "Users"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Datos del usuario"
          },
          "401": {
            "description": "No autorizado"
          }
        }
      },
      "put": {
        "summary": "Actualizar un usuario",
        "tags": [
          "Users"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string"
                  },
                  "role": {
                    "type": "string",
                    "enum": [
                      "admin",
                      "nurse"
                    ]
                  },
                  "is_active": {
                    "type": "boolean"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Usuario actualizado"
          },
          "401": {
            "description": "No autorizado"
          }
        }
      },
      "delete": {
        "summary": "Eliminar un usuario (Desactivar)",
        "tags": [
          "Users"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Usuario eliminado correctamente"
          },
          "401": {
            "description": "No autorizado"
          }
        }
      }
    }
  },
...
}